+++ KingAcadamey/KingAcadamey/Services/ArithmeticMeanService/IArithemticMeanService.cs
@@ -3,6 +3,6 @@ namespace KingAcadamey.Services.ArithmeticMeanService
 {
     public interface IArithemticMeanService
     {
-        double CalculateArithmeticMean(List<double> values);
+        decimal CalculateArithmeticMean(List<decimal> values);
     }
 }
